Move TokenRange
