+++ server/src/compiler_analyzer/functionCall.ts
@@ -6,7 +6,7 @@ import {SymbolScope} from "./symbolScope";
 import {applyTemplateTranslator, ResolvedType, TemplateTranslator} from "./resolvedType";
 import {analyzerDiagnostic} from "./analyzerDiagnostic";
 import {TokenObject} from "../compiler_tokenizer/tokenObject";
-import {TokenRange} from "../compiler_parser/tokenRange";
+import {TokenRange} from "../compiler_tokenizer/tokenRange";
 import {evaluateConversionCost} from "./typeConversion";
 
 interface FunctionCallArgs {
